Renamed DoAsync as Async
