+++ src/Command/Game/Async/LegalCommand.php
