+++ src/Command/Game/Mode/AbstractMode.php
@@ -5,8 +5,8 @@ namespace ChessServer\Command\Game\Mode;
 use ChessServer\Command\Game\Game;
 use ChessServer\Command\Game\Async\PlayCommand;
 use ChessServer\Command\Game\Async\PlayLanCommand;
-use ChessServer\Command\Game\DoAsync\LegalCommand;
-use ChessServer\Command\Game\DoAsync\UndoCommand;
+use ChessServer\Command\Game\Async\LegalCommand;
+use ChessServer\Command\Game\Async\UndoCommand;
 use Firebase\JWT\JWT;
 use Firebase\JWT\Key;
